mkfs.btrfs on ARM
Hi,

    I tried to run the mkfs.btrfs command on an AT91SAM9260 processor
(ARM, at91).
The cross compiling was quite simple but it doesn't work on the target
platform.
The mkfs.btrfs stopped with an assert and the filesystem wasn't mountable.

    I looked for solution on the internet. Some people noticed this
issue too, but nobody has a patch about it.
I found an interesting information, that some ARM processor isn't able
to access a word from an
halfword aligned location
(http://www.aleph1.co.uk/chapter-10-arm-structured-alignment-faq).
The gcc consider to use the correct alignment but in some situation the
problem is remaining.

    In the btrfs-progs there are a struct named extent_buffer. It has a
data field which could contains a bunch
of unaligned structs. The gcc can't access the member fields of these
structs correctly.
I found a workaround/solution that solve this problem. If I use the
memcpy command instead of setting or getting
members directly, the problem is eliminated.

    I have written a patch which works correctly on my platform.
I have attached this patch to my letter.

Best regards,
    Csaba Tóth, Watt 22 Ltd.